feature: prepare for add mmlu_evaluator
