+++ mlora/utils.py
@@ -1,9 +1,41 @@
 from typing import Dict
 from mlora.model_llama import LlamaModel
 from transformers import LlamaForCausalLM
+
 import os
 import json
 import torch
+import random
+import logging
+
+
+def setup_seed(seed):
+    torch.manual_seed(seed)
+    torch.cuda.manual_seed_all(seed)
+    random.seed(seed)
+
+
+def setup_logging(log_level: str = "INFO", log_file: str = None):
+    # set the logger
+    log_handlers = [logging.StreamHandler()]
+    if log_file is not None:
+        log_handlers.append(logging.FileHandler(log_file))
+
+    logging.basicConfig(format="[%(asctime)s] m-LoRA: %(message)s",
+                        level=log_level,
+                        handlers=log_handlers,
+                        force=True)
+
+
+def setup_cuda_check():
+    # check the enviroment
+    if torch.cuda.is_available():
+        logging.info('NVIDIA CUDA initialized successfully.')
+        logging.info('Total %i GPU(s) detected.' % torch.cuda.device_count())
+    else:
+        logging.error(
+            'm-LoRA requires NVIDIA CUDA computing capacity. Please check your PyTorch installation.')
+        exit(1)
 
 
 def convert_hf_to_pth(source: str, dest: str):
