+++ requirements.txt
@@ -7,10 +7,6 @@ bitsandbytes==0.41.1
 sentencepiece==0.1.99
 scipy==1.10.1
 xformers==0.0.21
-nltk
-jieba
-rouge
-rouge_chinese
 flask
 peft
 protobuf==3.20.2
